# feat: reuse public matrix for preimage