--- a/src/poly/dcrt/sampler/trapdoor.rs
+++ b/src/poly/dcrt/sampler/trapdoor.rs
@@ -3,7 +3,7 @@ use rayon::prelude::*;
 use std::sync::Arc;
 
 use crate::{
-    parallel_chunk_iter, parallel_iter,
+    parallel_iter,
     poly::{
         dcrt::{DCRTPoly, DCRTPolyMatrix, DCRTPolyParams},
         sampler::PolyTrapdoorSampler,
@@ -224,12 +224,13 @@ impl PolyTrapdoorSampler for DCRTPolyTrapdoorSampler {
 
         debug_mem("SetMatrixElement public_matrix_ptr completed");
 
-        let preimages: Vec<_> = (0..num_block)
+        let preimages: Vec<_> = parallel_iter!(0..num_block)
             .map(|i| {
                 let start_col = i * size;
                 let end_col = (start_col + size).min(target_cols);
                 let target_block = target.slice(0, size, start_col, end_col);
-                debug_mem(format!("preimage iter: start_col = {}", start_col));
+                debug_mem(format!("preimage iter : start_col = {}", start_col));
+
                 self.process_preimage_block(params, trapdoor, &public_matrix, &target_block, size)
             })
             .collect();