## Description

Don't duplicate public matrix generation per each loop of the preimage sample step.

## perf

It doesn't have a memory advantage, but it has slight time advantage + nonduplicating logic

But this advantage becomes huge, especially if we were to use mmap(because if then `SetElement` takes too long).

before
```
2025-03-27T03:24:57.179890Z  INFO test_io_dummy_param::test: Time to obfuscate: 77.02023275s
2025-03-27T03:25:25.357265Z  INFO test_io_dummy_param::test: [false, false, false, true]
2025-03-27T03:25:25.357277Z  INFO test_io_dummy_param::test: Time for evaluation: 28.177009917s
2025-03-27T03:25:25.357279Z  INFO test_io_dummy_param::test: Total time: 105.197242667s
```

after

```
2025-03-27T03:37:40.327227Z  INFO test_io_dummy_param::test: Time to obfuscate: 74.563497625s
2025-03-27T03:38:07.560632Z  INFO test_io_dummy_param::test: [true, true, false, false]
2025-03-27T03:38:07.560641Z  INFO test_io_dummy_param::test: Time for evaluation: 27.233052833s
2025-03-27T03:38:07.560643Z  INFO test_io_dummy_param::test: Total time: 101.796550458s
```